renamed to types.ts
